Small changes made to file.
Reference added, whitespace removed, second particle included, filename changed.

A 2x2 grid is sufficient as the grid has no spatial dependence so is uniform at any given timestep. This reduces the grid file size.
